+++ src/main/kotlin/org/rust/ide/annotator/RustHighlightingAnnotator.kt
@@ -9,37 +9,34 @@ import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.impl.mixin.isMut
 
 class RustHighlightingAnnotator : Annotator {
-    override fun annotate(element: PsiElement, holder: AnnotationHolder) =
-        element.accept(RustHighlightingAnnotatorVisitor(holder))
-}
-
-private class RustHighlightingAnnotatorVisitor(private val holder: AnnotationHolder) : RustVisitor() {
-    override fun visitAttr(o: RustAttr) {
-        holder.highlight(o, RustColors.ATTRIBUTE)
-    }
-
-    override fun visitMacroExpr(o: RustMacroExpr) {
-        holder.highlight(o.identifier, RustColors.MACRO)
-        holder.highlight(o.excl, RustColors.MACRO)
-    }
-
-    override fun visitTypeParam(o: RustTypeParam) {
-        holder.highlight(o, RustColors.TYPE_PARAMETER)
-    }
-
-    override fun visitPatBinding(o: RustPatBinding) {
-        if (o.isMut) {
-            holder.highlight(o.identifier, RustColors.MUT_BINDING)
+    override fun annotate(element: PsiElement, holder: AnnotationHolder) = element.accept(object : RustVisitor() {
+        override fun visitAttr(o: RustAttr) {
+            holder.highlight(o, RustColors.ATTRIBUTE)
         }
-    }
 
-    override fun visitPathPart(o: RustPathPart) {
-        o.reference.resolve().let {
-            if (it is RustPatBinding && it.isMut) {
+        override fun visitMacroExpr(o: RustMacroExpr) {
+            holder.highlight(o.identifier, RustColors.MACRO)
+            holder.highlight(o.excl, RustColors.MACRO)
+        }
+
+        override fun visitTypeParam(o: RustTypeParam) {
+            holder.highlight(o, RustColors.TYPE_PARAMETER)
+        }
+
+        override fun visitPatBinding(o: RustPatBinding) {
+            if (o.isMut) {
                 holder.highlight(o.identifier, RustColors.MUT_BINDING)
             }
         }
-    }
+
+        override fun visitPathPart(o: RustPathPart) {
+            o.reference.resolve().let {
+                if (it is RustPatBinding && it.isMut) {
+                    holder.highlight(o.identifier, RustColors.MUT_BINDING)
+                }
+            }
+        }
+    })
 }
 
 private fun AnnotationHolder.highlight(element: PsiElement?, textAttributes: TextAttributesKey) {
