+++ src/main/kotlin/org/rust/lang/core/psi/visitors/RustVisitorEx.kt
@@ -1,4 +1,7 @@
-package org.rust.lang.core.psi
+package org.rust.lang.core.psi.visitors
+
+import org.rust.lang.core.psi.RustLiteral
+import org.rust.lang.core.psi.RustVisitor
 
 open class RustVisitorEx : RustVisitor() {
     open fun visitLiteral(literal: RustLiteral) = visitElement(literal)
